style: components안에 main & reservations 폴더 생성
